+++ scalr/resource_scalr_workspace_migrate_test.go
@@ -30,3 +30,56 @@ func TestResourceScalrWorkspaceStateUpgradeV0(t *testing.T) {
 		t.Fatalf("\n\nexpected:\n\n%#v\n\ngot:\n\n%#v\n\n", expected, actual)
 	}
 }
+
+func testResourceScalrWorkspaceStateDataV1VcsRepo() map[string]interface{} {
+	return map[string]interface{}{
+		"id": "my-org/test",
+		"vcs_repo": map[string]interface{}{
+			"oauth_token_id": "test_provider_id",
+			"identifier":     "test_identifier",
+		},
+	}
+}
+
+func testResourceScalrWorkspaceStateDataV2() map[string]interface{} {
+	v1 := testResourceScalrWorkspaceStateDataV1VcsRepo()
+	vcsRepo := v1["vcs_repo"].(map[string]interface{})
+	return map[string]interface{}{
+		"id": v1["id"],
+		"vcs_repo": map[string]interface{}{
+			"identifier": vcsRepo["identifier"],
+		},
+		"vcs_provider_id": vcsRepo["oauth_token_id"],
+	}
+}
+
+func testResourceScalrWorkspaceStateDataV2NoVcs() map[string]interface{} {
+	v1 := testResourceScalrWorkspaceStateDataV1()
+	return map[string]interface{}{
+		"id": v1["id"],
+	}
+}
+
+func TestResourceScalrWorkspaceStateUpgradeV1(t *testing.T) {
+	expected := testResourceScalrWorkspaceStateDataV2()
+	actual, err := resourceScalrWorkspaceStateUpgradeV1(testResourceScalrWorkspaceStateDataV1VcsRepo(), nil)
+	if err != nil {
+		t.Fatalf("error migrating state: %s", err)
+	}
+
+	if !reflect.DeepEqual(expected, actual) {
+		t.Fatalf("\n\nexpected:\n\n%#v\n\ngot:\n\n%#v\n\n", expected, actual)
+	}
+}
+
+func TestResourceScalrWorkspaceStateUpgradeV1NoVcs(t *testing.T) {
+	expected := testResourceScalrWorkspaceStateDataV2NoVcs()
+	actual, err := resourceScalrWorkspaceStateUpgradeV1(testResourceScalrWorkspaceStateDataV1(), nil)
+	if err != nil {
+		t.Fatalf("error migrating state: %s", err)
+	}
+
+	if !reflect.DeepEqual(expected, actual) {
+		t.Fatalf("\n\nexpected:\n\n%#v\n\ngot:\n\n%#v\n\n", expected, actual)
+	}
+}